refactor: move websocket/config.go to new package to avoid import loops and add mutex and getters/setters
